Removed the correct number of input note check for the constructAtomicSwap() proof.
It does not add additional security (the proof is constructed on the client side) and it also  prevents testing that the verification contract throws when an incorrect num. of notes are provided.

Also added a constructIncorrectAtomicSwap() method. This deliberately constructs a proof for an incorrect number of input notes. The output is used later to check the verification contract fails for an incorrect number of notes